Remove type_info field from plist
When moving type_information from plist to qos, the field in struct
plist accidentally was not removed. This commit removes the field
and fixes sedp alive handling so that it uses the type identifier
from qos. Referencing the typelookup meta entry is moved to
proxy endpoint creation where the type id is actually stored in
the endpoint.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>