Add default_extensibility option to idlc
Add a command line option to the C generator for IDLC to set the
default extensibility for aggregated types. Aggregated types that
do not have their extensibility explicitly set, get the default
value. In case no default is provided, 'final' is used. Because
this is different than the default in the XTypes spec (which
is 'appendable'), a warning is shown in case there are types in
the IDL that do not have an explicit extensibility set and no
default is provided by the command line option.

Note that enumerated types also can have an extensibility annotation,
but that is currently not supported by the C generator.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>