Unalias declarators with aliased types correctly
Aliased struct and union member types are not correctly handled
in the idl c generator, resulting in incorrect serializer
instructions. This commit fixes this issue by adding the
IDL_VISIT_UNALIAS_TYPE_SPEC flag in the return value for
specific emit functions, so that the type will be unaliased
in the visit function.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>