Add default_extensibility option to idlc
Add a command line option to IDLC to set the default extensibility
for aggregated (and also for enumerated types when extensibility for these
types is added). Types that do not have their extensibility explicitly set,
get the provided default extensibility. In case no default is provided,
'final' is used. Because this is different than the default in the XTypes
spec (which is 'appendable'), a warning is shown in case there are aggregated
types in the IDL that do not have an explicit extensibility set and no
default is provided by the command line option.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>