Refactor idlc parser and generator config
Some refactoring of the IDL compiler config for parser and generators.
The parser config is moved to a struct 'config' in pstate, to separate
config from state for the parser (more settings will be added e.g. by
the default-extensibility option).

For generators, config is moved from pstate to a separate parameter of the
generate function, so that pstate includes only config for IDL parsing and
does not include generator specific config.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>