Type object to topic descriptor transformation
This adds the implementation for generating Cyclone topic descriptors
from XTypes type meta-data. Using type identifiers that are included in
endpoint and topic discovery data (in the form of a type information
object), the type meta-data, which is a set of type objects for the type
identifiers and its dependencies, can be retrieved using the type
lookup service.

The function ddsi_topic_descriptor_from_type that is added in this
commit, implements the transformation from a ddsi_type (that wraps
a XTypes type) to a dds_topic_descriptor_t, that includes the type
meta data required by the cdrstream serializer to read and write
serialized data for this type.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>