Public API changes for find_topic
Update the public API for finding a topic or resolving a topic
descriptor:
- Refactor the existing dds_find_topic function (that has been
deprecated for some time) so that it can be used for find a topic
by its name and type information (from discovery data). This
replaces dds_find_topic_scoped, which is deprecated now.
- Replace dds_resolve_type by dds_create_topic_descriptor, which
allows to get a topic descriptor by a type_info object, optionally
waiting for the type to become available.
- dds_builtintopic_get_endpoint_typeid is changed so that it returns
the type_information and not only a single type identifier.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>