Remove sertype ref from ddsi_type and proxy_endpoint_common
In the former cyclone specific implementation of type discovery,
a sertype was created at the moment a remote type was resolved. To
keep the refcount for this sertype > 0, the proxy endpoints that
were using it had a reference to this sertype. With the implementation
of the XTypes based type discovery, the type meta-data (type ids
and objects) are stored in the type library, and no sertype is
created upon receiving complete type meta-data. This commit
removes the reference to a sertype from proxy_endpoint_common.

In addition, the sertype ref is also removed from ddsi_type, which
was there to get a type information object for a type when sending
an SEDP message. We can now use the type meta-data stored in the type
library to generate the type information object used in SEDP.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>