Move sertype default and serdata default to DDSC
This moves the default implementation for sertype and serdata (used for
application data) to the DDSC layer. Because this sertype implementation
was also used in DDSI internally for the type lookup endpoints, a new
sertype_cdr implementation is introduced. This implementation does not
support shared memory transport and supports no keys, as this is not
required for these endpoints.

With moving serdata_default, the serdatapool also moved to the DDSC layer,
from domaingv to the dds_domain object. Additionally, sertype_default in
the cdrstream interface is replaced by struct dds_cdrstream_desc (renamed
and moved ddsi_sertype_default_desc to cdrstream).

Signed-off-by: Dennis Potman <dennis@zettascale.tech>