Unreliable readers are always "in sync" with pwr
The "in sync" vs "out sync" exists to handle retrieving with historical
data and determining when the "wait for historical data" operation can
return, and to handle cases where a reader deliberately falls behind a
writer because of resource limits.

Neither case exists for unreliable (a.k.a. best-effort) readers. So
arguably, all unreliable readers should always state they are "in sync".

The difference is negligible in the case of *unreliable* remote writers,
but not in the case of *reliable* remote writers. In the latter it can
cause message loss.

Signed-off-by: Erik Boasson <eb@ilities.com>