Fix key serialization
Use correct key field order in serialized key-only sample. Update
the CDR stream serializer for this, and change the IDL compiler
and dynamic type builder to include the keys in definition order.
The sort-order (3rd column) in the key list of the topic descriptor
reflects the member-ID order of the key fields, which is used in
serialization of the key for calculating the key hash.

The calculation of the key flags is moved to the CDR stream
serializer and is done runtime, so that this can be used for both
IDLC and type builder generated descriptors.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>