Fix psmx metadata cdr_identifier usage
The cdr_identifier field in psmx meta-data should have a
valid encapsulation identifier, not just the XCDR version
(or undefined in case of a raw sample). This commit introduces
a vendor specific RTPS encapsulation kind for a raw sample
in native endianess, and it fixes the use of the
cdr_identifier field.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>