Return a loaned_sample from dds_loan_shared_memory_buffer
Update (and rename) dds_loan_shared_memory_buffer so that it
returns a dds_loaned_sample_t and not just a raw loan pointer
(that can't be used with dds_write because the buffer is not
registered in the loan administration). With this change, the
loaned sample (with custom size) can be used with the regular
write function; only difference with the regular request_loan
is that the user can specify the size in this case.

The request_raw_loan fn prototype is removed from the PSMX
endpoint ops; request_loan with the user-specified size
is used instead.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>