Fix number of ops in descriptor and find member ID table
Updates IDLC and typebuilder so that the number of ops in a topic
descriptor is the actual number of instructions in the ops parameter.
This allows to find the member ID table in the ops in cdrstream,
without the need for an offset to this table in the descriptor
(the output of older versions of IDLC is still handled correctly)

Initializing a CDR stream descriptor now needs the ops count as
a parameter, so a new variant dds_cdrstream_desc_init_with_nops
is introduced and the old variant is left unchanged.

Signed-off-by: Dennis Potman <dennis@zettascale.tech>